src/pyproject.toml
[tool.poetry]
name = "pizzeria-new"
version = "0.1.0"
description = ""
authors = ["Bpem9 <itimeq@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
pydantic = "^1.10.8"
gunicorn = "^20.1.0"
setuptools = "^67.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
